tests: move in here 'fd_dj_accounts/tests'
(overwrite the existing dir contents)
